# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Netflix Clone API",
        "description": "API Documentation for Netflix Clone Video Platform",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get list of videos",
                "description": "Display a listing of the resource.",
                "operationId": "2090c29a8b2a49778e2de02f54ec7606",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter videos by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for videos",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of videos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Video"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "Upload a new video",
                "description": "Store a newly created resource in storage.",
                "operationId": "904069e3a71065b7227cc44e7dc50fb2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreVideoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Video"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Video created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreVideoRequest": {
                "title": "Store Video Request",
                "description": "Request body for creating a new video",
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Introduction to Laravel"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Learn the basics of Laravel framework"
                    },
                    "url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/videos/laravel-intro.mp4"
                    },
                    "thumbnail": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/thumbnails/laravel-intro.jpg"
                    },
                    "category_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    "video": {
                        "description": "Video file to upload",
                        "type": "string",
                        "format": "binary"
                    },
                    "thumbnail_file": {
                        "description": "Thumbnail image to upload",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Category model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Programming"
                    },
                    "description": {
                        "type": "string",
                        "example": "Programming tutorials and guides"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Video": {
                "title": "Video",
                "description": "Video model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduction to Laravel"
                    },
                    "description": {
                        "type": "string",
                        "example": "Learn the basics of Laravel framework"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/videos/laravel-intro.mp4"
                    },
                    "thumbnail": {
                        "type": "string",
                        "example": "https://example.com/thumbnails/laravel-intro.jpg"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "published",
                            "failed"
                        ],
                        "example": "published"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}